+++ src/main/java/com/ujax/application/mail/outbox/MailOutboxEventLogger.java
@@ -5,7 +5,6 @@ import java.time.LocalDateTime;
 import org.springframework.stereotype.Component;
 
 import com.ujax.domain.mail.MailOutbox;
-import com.ujax.domain.mail.MailOutboxEventType;
 import com.ujax.domain.mail.MailOutboxStatus;
 
 import lombok.extern.slf4j.Slf4j;
@@ -14,13 +13,20 @@ import lombok.extern.slf4j.Slf4j;
 @Component
 public class MailOutboxEventLogger {
 
+	private static final String EVENT_ENQUEUED = "ENQUEUED";
+	private static final String EVENT_PROCESSING_STARTED = "PROCESSING_STARTED";
+	private static final String EVENT_RECOVERED = "RECOVERED";
+	private static final String EVENT_SENT = "SENT";
+	private static final String EVENT_RETRY_SCHEDULED = "RETRY_SCHEDULED";
+	private static final String EVENT_FAILED = "FAILED";
+
 	private static final String LOG_PATTERN =
 		"event=mail_outbox eventType={} outboxId={} mailType={} recipientEmail={} fromStatus={} toStatus={} attemptNo={} nextAttemptAt={} sentAt={} lastError={}";
 
 	public void logEnqueued(MailOutbox outbox) {
 		log.info(
 			LOG_PATTERN,
-			MailOutboxEventType.ENQUEUED,
+			EVENT_ENQUEUED,
 			outbox.getId(),
 			outbox.getMailType(),
 			outbox.getRecipientEmail(),
@@ -33,36 +39,47 @@ public class MailOutboxEventLogger {
 		);
 	}
 
-	public void logTransition(
-		MailOutbox outbox,
-		MailOutboxEventType eventType,
-		MailOutboxStatus fromStatus,
-		MailOutboxStatus toStatus
-	) {
-		if (eventType == MailOutboxEventType.RECOVERED || eventType == MailOutboxEventType.RETRY_SCHEDULED) {
-			log.warn(
-				LOG_PATTERN,
-				eventType,
-				outbox.getId(),
-				outbox.getMailType(),
-				outbox.getRecipientEmail(),
-				fromStatus,
-				toStatus,
-				outbox.getAttemptNo(),
-				outbox.getNextAttemptAt(),
-				null,
-				outbox.getLastError()
-			);
-			return;
-		}
+	public void logProcessingStarted(MailOutbox outbox, MailOutboxStatus fromStatus) {
 		log.info(
 			LOG_PATTERN,
-			eventType,
+			EVENT_PROCESSING_STARTED,
 			outbox.getId(),
 			outbox.getMailType(),
 			outbox.getRecipientEmail(),
 			fromStatus,
-			toStatus,
+			outbox.getStatus(),
+			outbox.getAttemptNo(),
+			outbox.getNextAttemptAt(),
+			null,
+			outbox.getLastError()
+		);
+	}
+
+	public void logRecovered(MailOutbox outbox) {
+		log.warn(
+			LOG_PATTERN,
+			EVENT_RECOVERED,
+			outbox.getId(),
+			outbox.getMailType(),
+			outbox.getRecipientEmail(),
+			MailOutboxStatus.PROCESSING,
+			outbox.getStatus(),
+			outbox.getAttemptNo(),
+			outbox.getNextAttemptAt(),
+			null,
+			outbox.getLastError()
+		);
+	}
+
+	public void logRetryScheduled(MailOutbox outbox, MailOutboxStatus fromStatus) {
+		log.warn(
+			LOG_PATTERN,
+			EVENT_RETRY_SCHEDULED,
+			outbox.getId(),
+			outbox.getMailType(),
+			outbox.getRecipientEmail(),
+			fromStatus,
+			outbox.getStatus(),
 			outbox.getAttemptNo(),
 			outbox.getNextAttemptAt(),
 			null,
@@ -73,7 +90,7 @@ public class MailOutboxEventLogger {
 	public void logSent(MailOutbox outbox, LocalDateTime sentAt) {
 		log.info(
 			LOG_PATTERN,
-			MailOutboxEventType.SENT,
+			EVENT_SENT,
 			outbox.getId(),
 			outbox.getMailType(),
 			outbox.getRecipientEmail(),
@@ -89,7 +106,7 @@ public class MailOutboxEventLogger {
 	public void logFailed(MailOutbox outbox, String lastError) {
 		log.warn(
 			LOG_PATTERN,
-			MailOutboxEventType.FAILED,
+			EVENT_FAILED,
 			outbox.getId(),
 			outbox.getMailType(),
 			outbox.getRecipientEmail(),
