+++ src/main/java/com/ujax/application/mail/outbox/MailSender.java
@@ -1,6 +1,6 @@
 package com.ujax.application.mail.outbox;
 
-import com.ujax.application.mail.RenderedMailContent;
+import com.ujax.application.mail.template.RenderedMailContent;
 
 public interface MailSender {
