--- src/main/java/com/ujax/application/mail/outbox/PreparedMailMessage.java
@@ -1,9 +1,0 @@
-package com.ujax.application.mail.outbox;
-
-import com.ujax.application.mail.RenderedMailContent;
-
-public record PreparedMailMessage(
-	String subject,
-	RenderedMailContent content
-) {
-}
