+++ src/main/java/com/ujax/application/mail/outbox/handler/MailOutboxHandler.java
@@ -1,5 +1,6 @@
-package com.ujax.application.mail.outbox;
+package com.ujax.application.mail.outbox.handler;
 
+import com.ujax.application.mail.outbox.message.PreparedMailMessage;
 import com.ujax.domain.mail.MailType;
 
 public interface MailOutboxHandler {
