+++ src/main/java/com/ujax/application/mail/outbox/handler/SignupVerificationMailOutboxHandler.java
@@ -1,4 +1,4 @@
-package com.ujax.application.mail.outbox;
+package com.ujax.application.mail.outbox.handler;
 
 import java.io.IOException;
 import java.util.Objects;
@@ -6,7 +6,9 @@ import java.util.Objects;
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.ujax.application.mail.UjaxMailTemplateRenderer;
+import com.ujax.application.mail.outbox.message.PreparedMailMessage;
+import com.ujax.application.mail.outbox.message.SignupVerificationMailPayload;
+import com.ujax.application.mail.template.SignupVerificationMailTemplateRenderer;
 import com.ujax.domain.mail.MailType;
 
 import lombok.RequiredArgsConstructor;
@@ -29,7 +31,7 @@ public class SignupVerificationMailOutboxHandler implements MailOutboxHandler {
 			String code = Objects.requireNonNull(payload.code(), "code must not be null");
 			return new PreparedMailMessage(
 				"[UJAX] 회원가입 인증 코드 - [ " + code + " ]",
-				UjaxMailTemplateRenderer.renderSignupVerification(code)
+				SignupVerificationMailTemplateRenderer.render(code)
 			);
 		} catch (IOException exception) {
 			throw new IllegalArgumentException("invalid signup verification mail payload", exception);
