+++ src/main/java/com/ujax/application/mail/outbox/message/PreparedMailMessage.java
@@ -1,0 +1,9 @@
+package com.ujax.application.mail.outbox.message;
+
+import com.ujax.application.mail.template.RenderedMailContent;
+
+public record PreparedMailMessage(
+	String subject,
+	RenderedMailContent content
+) {
+}
