+++ src/main/java/com/ujax/application/mail/outbox/message/WorkspaceInviteMailPayload.java
@@ -1,4 +1,4 @@
-package com.ujax.application.mail.outbox;
+package com.ujax.application.mail.outbox.message;
 
 public record WorkspaceInviteMailPayload(
 	Long workspaceId,
