+++ src/main/java/com/ujax/application/mail/template/RenderedMailContent.java
@@ -1,4 +1,4 @@
-package com.ujax.application.mail;
+package com.ujax.application.mail.template;
 
 public record RenderedMailContent(
 	String plainText,
