--- src/main/java/com/ujax/domain/mail/MailOutboxEventType.java
@@ -1,12 +1,0 @@
-package com.ujax.domain.mail;
-
-public enum MailOutboxEventType {
-
-	ENQUEUED,
-	PROCESSING_STARTED,
-	RECOVERED,
-	SENT,
-	RETRY_SCHEDULED,
-	FAILED
-
-}
